requirements.txt
# FastAPI相关依赖
fastapi==0.112.0
uvicorn==0.30.6

# 数据库相关依赖
sqlalchemy==2.0.31
psycopg2-binary==2.9.9

# 数据验证和模型
pydantic==2.8.2
pydantic-settings==2.3.4

# 安全相关\ypasslib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# 环境配置
python-dotenv==1.0.1

# 缓存管理
redis==5.0.6

# 开发工具
black==24.4.2
isort==5.13.2
flake8==7.1.0

# 文档生成
mkdocs==1.5.3
mkdocs-material==9.5.18

# 测试工具
pytest==8.2.2
pytest-asyncio==0.23.6
httpx==0.27.0

# 其他工具
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0